🚚 shop/05 ->shop/shop-content 파일 변경
sub-issues:
- [NAILCASE-423](https://nailcase.atlassian.net/browse/NAILCASE-423)
