+++ src/app/(customer)/shop/[shopId]/page.tsx
@@ -3,13 +3,12 @@
 import dynamic from "next/dynamic"
 import { useRouter } from "next/navigation"
 
-import { NTButton } from "@/component/common/atom/nt-button"
 import CustomerShopBanner from "@/component/custom/customer/shop/01"
 import { useShopById } from "@/hook/use-shop-controller"
 import { convertStringToInteger } from "@/util/common"
 
 const CustomerShopContent = dynamic(
-	() => import("@/component/custom/customer/shop/05/index"),
+	() => import("@/component/custom/customer/shop/shop-content/index"),
 	{ ssr: false },
 )
 
@@ -28,10 +27,6 @@ export default function CustomerShop({ params }: CustomerShopPT) {
 		isLoading,
 	} = useShopById(convertStringToInteger(shopId))
 
-	const onClickReservationButton = () => {
-		router.push(`/shop/${shopId}/reservation`)
-	}
-
 	if (isLoading) {
 		return (
 			<div className="flex h-screen items-center justify-center">
@@ -48,14 +43,7 @@ export default function CustomerShop({ params }: CustomerShopPT) {
 		return (
 			<div className="h-full w-full">
 				<CustomerShopBanner shopId={convertStringToInteger(params.shopId)} />
-				<div className="mt-[32px] flex w-full flex-col gap-[32px]">
-					<div className="flex w-full flex-col gap-[20px]">
-						<NTButton flexible="full" onClick={onClickReservationButton}>
-							예약하기
-						</NTButton>
-					</div>
-					<CustomerShopContent shopId={convertStringToInteger(params.shopId)} />
-				</div>
+				<CustomerShopContent shopId={convertStringToInteger(params.shopId)} />
 			</div>
 		)
 	} else {
